+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/DI.android.kt
@@ -1,6 +1,8 @@
 package de.connect2x.trixnity.messenger
 
+import de.connect2x.trixnity.messenger.media.AndroidAudioRecorder
 import de.connect2x.trixnity.messenger.media.AndroidMediaPlayer
+import de.connect2x.trixnity.messenger.media.AudioRecorder
 import de.connect2x.trixnity.messenger.media.MediaPlayer
 import de.connect2x.trixnity.messenger.multi.CopyMultiMessengerSingletons
 import de.connect2x.trixnity.messenger.util.ActivityGetter
@@ -20,5 +22,8 @@ actual fun platformModule(): Module = module {
     }
     single<MediaPlayer> {
         AndroidMediaPlayer(get(), get())
+    }
+    single<AudioRecorder> {
+        AndroidAudioRecorder(get(),get(), get(), get(), get())
     }
 }